Add javascript import/require folding
The tens of lines of imports at the beginning of every component drive
me nuts. This commit defines a pretty simple FoldFunc that assigns them
a level higher than the value of my foldlevelstart, so they are hidden
by default.

There is plenty more work to do here to improve folding of components.