Show "WIP" warning based on master branch upstream
E.g. when on a "wip_some_feature" branch, we want to show the "WIP"
prompt warning if there is a WIP commit between origin/master and the
tip of "wip_some_feature". We want to specify the master branch here,
because it's unlikely that "wip_some_feature" has an upstream branch.